code/netbox_testbed.py
import logging
import json
import hvac
import pynetbox
from dotenv import dotenv_values
from pyats.topology import loader
from templates import testbed_device_template

logger = logging.getLogger(__name__)

