requirements.txt
hvac==2.3.0
pynetbox==7.4.0
python-dotenv==1.0.1
